fix phpstan and tidy up Process class

+++ src/Console/ProgressBar.php
@@ -27,7 +27,7 @@ final readonly class ProgressBar
 
         $out = "<fg=$barColor;bg=gray>" .
             str_repeat(self::CHARS[8], $fullBlocks) .
-            ($partialBlock > 0 ? self::CHARS[round($partialBlock * 8)] : '') .
+            ($partialBlock > 0 ? self::CHARS[(int)round($partialBlock * 8)] : '') .
             str_repeat(' ', $width - $fullBlocks - ($partialBlock > 0 ? 1 : 0)) .
             '</>';
         if ($textColor !== null) {
